separates files in modules
