+++ robot/ImageAnalyzer.py
